Split file into regions (+ruff)
A few text editors support marking a file into foldable regions.
This #region ... #endregion syntax is from VS and VS code, but Vim can
be configured to support it as well.

These regions are a band-aid on what should probably be multiple files.
It does turn out that the file is logically grouped well-enough that
folding is useful, though (kudos!). If/when splitting happens it will
happen along these markers.